requirements.txt
torch>=2.0
torchvision>=0.15
pandas>=2.0
Pillow>=9.0
opencv-python>=4.6
onnx>=1.12
onnx-tf>=1.10
tensorflow==2.11.0
matplotlib>=3.5
pyyaml>=6.0
